requirements.txt
python-telegram-bot==20.3
PyMuPDF==1.22.5
pdf2image==1.16.3
python-docx==0.8.12
python-pptx==0.6.21
Pillow==10.0.0
reportlab==4.0.3
openpyxl==3.1.2
